requirements.txt
wykop-sdk-reborn>=0.2.1
markovify>=0.8.3